+++ src/js/shoppingCart.mjs
@@ -13,7 +13,6 @@ export default function shoppingCart() {
 
 function cartItemTemplate(item) {
   const newItem = `<li class="cart-card divider">
-  
   <a href="#" class="cart-card__image">
   <img
   src="${item.Images.PrimaryMedium}"
@@ -30,32 +29,4 @@ function cartItemTemplate(item) {
 </li>`;
 
   return newItem;
-}
-
-function removeFromLocalStorage(id) {
-  const cartItems = getLocalStorage("so-cart");
-
-  // search for id of item to remove from storage
-  cartItems.forEach((element, index) => {
-    if (element.Id === id) {
-      console.log(element.Id);
-      console.log(id);
-      cartItems.splice(index, 1);
-      console.log(cartItems);
-    }
-  });
-}
-
-export function removeCartListeners() {
-  const productList = document.querySelectorAll(".cart-card");
-
-  // loop through items to add event-listener for removing cart items
-  productList.forEach((product) => {
-    const span = product.querySelector(".remove-item");
-    span.addEventListener("click", (e) => {
-      // console.log(e.target.id);
-      // console.log("Removing item!");
-      removeFromLocalStorage(e.target.id);
-    });
-  });
 }
